+++ src/SegGeometryEpix100V1.py
@@ -354,9 +354,9 @@ class SegGeometryEpix100V1(SegGeometry) :
             mbits: +1 - mask edges
                    +2 - mask two central columns 
         """
-        zero_col = np.zeros(sp._rows,dtype=np.int)
-        zero_row = np.zeros(sp._cols,dtype=np.int)
-        mask     = np.ones((sp._rows,sp._cols),dtype=np.int)
+        zero_col = np.zeros(sp._rows,dtype=np.uint8)
+        zero_row = np.zeros(sp._cols,dtype=np.uint8)
+        mask     = np.ones((sp._rows,sp._cols),dtype=np.uint8)
 
         if mbits & 1 : 
         # mask edges
